Added initial docs, moved Cord19Article to collection package.
