+++ pyserini/collection/pycollection.py
@@ -14,6 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import json
 import logging
 import re
 
@@ -128,3 +129,59 @@ class SourceDocument:
         self.indexable = self.object.indexable()
         self.contents = self.object.contents()
         self.raw = self.object.raw()
+
+
+class Cord19Article:
+    """Wrapper class for a raw JSON article from AI2's COVID-19 Open Research Dataset (CORD-19).
+
+    Parameters
+    ----------
+    document : JDocument
+        Underlying Lucene ``Document``.
+    """
+
+    def __init__(self, doc):
+        self.json = json.loads(doc)
+        # Perform some basic error checking, throw an exception if user tries to instantiate with an underlying
+        # Lucene document that isn't from CORD-19
+        if 'cord_uid' in self.json:
+            self.full_text = False
+        elif 'paper_id' in self.json:
+            self.full_text = True
+        else:
+            raise TypeError
+
+    def is_full_text(self):
+        return self.json['has_full_text']
+
+    def cord_uid(self):
+        return self.json['cord_uid']
+
+    def title(self):
+        try:
+            if self.is_full_text():
+                return self.json['metadata']['title']
+            else:
+                return self.json['csv_metadata']['title']
+        except KeyError:
+            return ''
+
+    def abstract(self):
+        try:
+            # For a full-text article, we can grab the abstract from two independent sources, the metadata or the
+            # actual full text. Here, we make the decision to use the metadata, even for full text.
+            return self.json['csv_metadata']['abstract']
+        except KeyError:
+            return ''
+
+    def metadata(self):
+        return self.json['csv_metadata']
+
+    def body(self):
+        try:
+            if self.is_full_text():
+                return [entry['text'] for entry in self.json['body_text']]
+            else:
+                return []
+        except KeyError:
+            return ''
